requirements.txt
streamlit>=1.33
langgraph>=0.1.7
langchain>=0.1.16
requests>=2.31
chromadb>=0.6.1
sentence-transformers>=2.2
jinja2>=3.1
openai>=1.30
openai-whisper==20250625
google-generativeai>=0.7
google-api-python-client>=2.127
google-auth>=2.29
sounddevice==0.5.2
numpy>=1.26,<2.0
scipy>=1.14
pydub==0.25.1
google-genai==1.38.0